refactor: unify database types and payment status
