+++ src/lib/database/types.ts
@@ -1,6 +1,7 @@
 
 import { PostgrestError } from '@supabase/supabase-js';
 import { Database } from '@/types/database.types';
+export * from '@/types/database-common';
 
 // Helper type for easy table access
 export type Tables = Database['public']['Tables'];
@@ -40,16 +41,3 @@ export function isValidStatus<T extends { status: string }>(record: T, status: T
   return record.status === status;
 }
 
-// Export commonly used table types
-export type LeaseRow = TableRow<'leases'>;
-export type PaymentRow = TableRow<'unified_payments'>;
-export type VehicleRow = TableRow<'vehicles'>;
-export type ProfileRow = TableRow<'profiles'>;
-export type TrafficFineRow = TableRow<'traffic_fines'>;
-export type LegalCaseRow = TableRow<'legal_cases'>;
-
-// Common status types
-export type VehicleStatus = VehicleRow['status']; 
-export type LeaseStatus = LeaseRow['status'];
-export type PaymentStatus = PaymentRow['status']; 
-export type ProfileStatus = ProfileRow['status'];
