+++ src/types/database-common.ts
@@ -1,5 +1,6 @@
 
 import { Database } from '@/types/database.types';
+export { isSuccessResponse } from '@/lib/database/validation/typeGuards';
 
 // Core database types with consistent naming
 export type DatabaseSchema = Database['public'];
@@ -19,7 +20,13 @@ export type AgreementId = LeaseId; // Maintaining legacy compatibility
 // Standardized status types from database schema
 export type LeaseStatus = DatabaseTables['leases']['Row']['status'];
 export type VehicleStatus = DatabaseTables['vehicles']['Row']['status'];
-export type PaymentStatus = string; // Using string since it could have different values
+export type PaymentStatus =
+  | 'pending'
+  | 'completed'
+  | 'overdue'
+  | 'cancelled'
+  | 'partially_paid'
+  | 'voided';
 export type ProfileStatus = string;
 export type MaintenanceStatus = string;
 
@@ -108,10 +115,6 @@ export type ListResponse<T> = DatabaseResponse<T[]>;
 export type SingleResponse<T> = DatabaseResponse<T>;
 
 // Helper type guards with consistent naming
-export function isSuccessResponse<T>(response: any): response is { data: T; error: null } {
-  return !response?.error && response?.data !== null;
-}
-
 export function hasEntityData<T>(response: any): response is { data: T; error: null } {
   return !response?.error && response?.data !== null;
 }
